fix: Remove api prefix from package
